Implemented the login functionality
