+++ lib/Constants.js
@@ -1,0 +1,8 @@
+export const actions = {
+    LOGIN: "Login",
+    LOGOUT: "Logout"
+}
+
+export const values = {
+    USER: "user",
+}
