+++ pages/users/[user]/dashboard.js
@@ -1,9 +1,22 @@
-const Dashboard = ()=>{
+const Dashboard = () => {
     return (
         <>
-            Dashboard
+            <button onClick={() => {
+
+            }}>
+                Dashboard
+            </button>
         </>
     )
 }
 
 export default Dashboard;
+
+export const getServerSideProps = async () => {
+
+    return {
+        props: {
+            mate: "fish"
+        },
+    }
+}
